Refactored MapView and updated layout
