+++ src/components/user-reservation/MapSideBar.jsx
@@ -1,16 +1,14 @@
-import { useState } from "react";
 import { getNext8Days } from "./DaysDropdown";
 import { useSelector, useDispatch } from "react-redux";
 import {
   setDate,
+  setSearch,
   fetchCoordinates,
   fetchClosestGarages,
   setPage,
-  setSearch,
 } from "../../utils/slice/reservationSlice";
 
-const SearchBar = () => {
-  // const [address, setAddress] = useState(null);
+const MapSideBar = () => {
   const { reservation } = useSelector((state) => state.reservation);
   const dispatch = useDispatch();
 
@@ -26,38 +24,43 @@ const SearchBar = () => {
   };
 
   return (
-    <form onSubmit={handleSubmit} className="w-3/4">
-      <div className="join flex justify-center w-full">
+    <form
+      onSubmit={handleSubmit}
+      className="flex flex-col items-center w-[90%]"
+    >
+      <div className="flex flex-col justify-center pb-3 w-full">
+        <label htmlFor="search">Book Parking Near</label>
         <input
           required
-          className="input input-bordered join-item w-full"
+          className="input input-bordered w-full"
           placeholder="Address, Place, City or Venue"
           type="text"
+          value={reservation.search}
           onChange={(e) => {
             dispatch(setSearch(e.target.value));
           }}
         />
+      </div>
+      <div className="flex flex-col pb-5 w-full">
+        <label htmlFor="search">Date</label>
         <select
           required
-          className="select select-bordered join-item"
+          className="select select-bordered "
           value={reservation.date || ""}
           onChange={(e) => {
             dispatch(setDate(e.target.value.replace(/\//g, "-")));
           }}
         >
-          <option value="" disabled>
-            Date
-          </option>
           {getNext8Days()}
         </select>
-        <div className="indicator">
-          <button type="submit" className="btn join-item">
-            Search
-          </button>
-        </div>
+      </div>
+      <div className="indicator flex justify-center items-center w-full">
+        <button type="submit" className="btn">
+          Update Search
+        </button>
       </div>
     </form>
   );
 };
 
-export default SearchBar;
+export default MapSideBar;
