+++ src/components/user-reservation/MapView.jsx
@@ -1,87 +1,15 @@
-import { useSelector, useDispatch } from "react-redux";
-import {
-  setGarageId,
-  setSelectedGarage,
-  fetchReservations,
-} from "../../utils/slice/reservationSlice";
-import TimeSlotList from "./TimeSlotList";
-import {
-  APIProvider,
-  Map,
-  Marker,
-  InfoWindow,
-} from "@vis.gl/react-google-maps";
-const apiKey = import.meta.env.VITE_GOOGLE_API_KEY;
+import MapSideBar from "./MapSideBar";
+import GoogleMap from "./GoogleMap";
 
 const MapView = () => {
-  const { mapCenter, closestGarages, reservationsList, selectedGarage } =
-    useSelector((state) => state.reservation);
-  const dispatch = useDispatch();
-
-  const createMarkers = (garageList) => {
-    return garageList.map((garage) => {
-      return (
-        <Marker
-          key={garage.lat + garage.lng}
-          position={{ lat: garage.lat, lng: garage.lng }}
-          onClick={() => dispatch(setSelectedGarage(garage))}
-        ></Marker>
-      );
-    });
-  };
-
-  const getResevations = async (garageId) => {
-    try {
-      await dispatch(setGarageId(garageId));
-      await dispatch(fetchReservations(garageId));
-    } catch (err) {
-      console.log(err);
-    }
-  };
-
   return (
-    mapCenter &&
-    closestGarages && (
-      <div>
-        <div className="flex justify-center">
-          <APIProvider apiKey={apiKey}>
-            <div style={{ height: "50vh", width: "75vw" }}>
-              <Map center={mapCenter} zoom={12}>
-                {/* WILL UPDATE STYLE SO THAT CENTER & GARAGE MARKERS ARE DISTINCT */}
-                <Marker position={mapCenter} />
-                {createMarkers(closestGarages)}
-                {selectedGarage && (
-                  <InfoWindow
-                    position={{
-                      lat: selectedGarage.lat,
-                      lng: selectedGarage.lng,
-                    }}
-                    onCloseClick={() => {
-                      dispatch(setSelectedGarage(false));
-                    }}
-                  >
-                    <h1>{selectedGarage.name || "GARAGE NAME"}</h1>
-                    <p>{selectedGarage.address || "ADDRESS"}</p>
-                    <p>
-                      {selectedGarage.city || "CITY"},{" "}
-                      {selectedGarage.state || "STATE"}
-                    </p>
-                    <p>{selectedGarage.distance || "1.5"} miles away</p>
-                    <button
-                      className="border-2 border-burgandy-p"
-                      onClick={() => getResevations(selectedGarage.id)}
-                    >
-                      Show Times
-                    </button>
-                  </InfoWindow>
-                )}
-              </Map>
-            </div>
-          </APIProvider>
-        </div>
-        {reservationsList && <TimeSlotList />}
+    <div className="flex flex-row w-full">
+      <div className="flex flex-col w-full items-center">
+        <h3 className="text-xl pb-5">Reserve your spot!</h3>
+        <MapSideBar />
       </div>
-    )
+      <GoogleMap />
+    </div>
   );
 };
